added further tests and included checker in scan datasets
