+++ tests/test_scan/test_scan_dataset.py
@@ -15,8 +15,6 @@ def setup_module(module):
 def test_corrupt_json_file():
     """ Tests what happens when a JSON file exists but is incomplete due to an issue encoding."""
     ds_id = ["c3s-cordex.output.EUR-11.IPSL.MOHC-HadGEM2-ES.rcp85.r1i1p1.IPSL-WRF381P.v1.day.psl.v20190212"]
-    scan_datasets(project='c3s-cordex', ds_ids=ds_id, paths=options.project_base_dirs['c3s-cordex'],
-                  mode='quick', location='ceda')
     try:
         scan_datasets(project='c3s-cordex', ds_ids=ds_id, paths=options.project_base_dirs['c3s-cordex'],
                       mode='quick', location='ceda')
@@ -35,6 +33,11 @@ def test_fake_corrupt_json_file(tmpdir):
     except json.decoder.JSONDecodeError as exc:
         print('[INFO] Corrupt JSON file found.')
         pass
+
+
+def test_file_checker():
+    scan_datasets(project='c3s-cordex', ds_ids=ds_id, paths=options.project_base_dirs['c3s-cordex'],
+                  mode='quick', location='ceda')
 
 
 def teardown_module(module):
